+++ src/testing/platform/testing-constants.ts
@@ -3,7 +3,7 @@ import type * as d from '@stencil/core/internal';
 import { QueuedLoadModule } from './load-module';
 
 export const styles: d.StyleMap = new Map();
-export const modeResolutionChain: d.ResolutionHandler[] = [];
+export const modeResolver: d.ResolutionHandler[] = [];
 
 /**
  * A mapping of custom element tags (e.g. `my-component`) to their constructor